+++ templates/rust-js/app/app.js
@@ -8,4 +8,4 @@ t.get("/rust").action("rust_hello") // This route uses a rust action
 
 t.get("/").reply("Ready to land on Titan Planet 🚀");
 
-t.start(3000, "Titan Running!");
+t.start(5100, "Titan Running!");