+++ titanpl-sdk/templates/app/app.js
@@ -7,4 +7,4 @@ t.post("/hello").action("hello") // pass a json payload { "name": "titan" }
 
 t.get("/").reply("Ready to land on Titan Planet 🚀");
 
-t.start(3000, "Titan Running!");
+t.start(5100, "Titan Running!");